refactor project to use AbstractInteractiveElement
